backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
opencv-python>=4.8.0
Pillow>=10.0.0
pyyaml>=6.0.0
python-multipart>=0.0.9
